gorkins/requirements.txt
tweepy==4.14.0
python-dotenv==1.0.0
requests==2.31.0
redis==5.0.1
flask==3.0.0
requests-oauthlib==1.3.1 